feat: 内容 Agent、多岗位知识库 RAG 与公众号样式还原
落地内容运营 Agent 任务流、按岗位 Agent/知识库管理（V48/V49），并完善公众号导入异步与正文样式还原；同步生产部署脚本与后台版本号构建注入。

+++ backend/src/main/java/com/miniprogram/service/impl/ContentInteractServiceImpl.java
@@ -58,7 +58,7 @@ public class ContentInteractServiceImpl implements ContentInteractService {
             dto.setLiked(false);
             dto.setFavorited(false);
         }
-        dto.setCommentCount(commentCount(contentId));
+        dto.setCommentCount((int) commentCount(contentId));
         return dto;
     }
